--- a/command.go
+++ b/command.go
@@ -28,7 +28,6 @@ const (
 	CmdStop
 	CmdQuit
 	CmdStatus
-	CmdLog
 )
 
 func (cmd CmdType) String() string {
@@ -53,8 +52,6 @@ func (cmd CmdType) String() string {
 		return "/quit"
 	case CmdStatus:
 		return "/status"
-	case CmdLog:
-		return "/log"
 	default:
 		panic(xerrors.Errorf("invalid cmd-type %d", byte(cmd)))
 	}
@@ -71,7 +68,6 @@ var cmdNames = [...][]byte{
 	CmdStop:    []byte(CmdStop.String()),
 	CmdQuit:    []byte(CmdQuit.String()),
 	CmdStatus:  []byte(CmdStatus.String()),
-	CmdLog:     []byte(CmdLog.String()),
 }
 
 func cmdTypeToPath(cmd CmdType) []byte {
@@ -354,47 +350,6 @@ func (cmd *StatusCmd) UnmarshalTDAQ(p []byte) error {
 	return dec.err
 }
 
-type LogCmd struct {
-	Name string // name of the TDAQ process
-	Addr string // address of log server
-}
-
-func newLogCmd(frame Frame) (LogCmd, error) {
-	var (
-		cmd LogCmd
-		err error
-	)
-
-	raw, err := cmdFrom(frame)
-	if err != nil {
-		return cmd, xerrors.Errorf("not a /log cmd: %w", err)
-	}
-
-	if raw.Type != CmdLog {
-		return cmd, xerrors.Errorf("not a /log cmd")
-	}
-
-	err = cmd.UnmarshalTDAQ(raw.Body)
-	return cmd, err
-}
-
-func (cmd LogCmd) CmdType() CmdType { return CmdLog }
-
-func (cmd LogCmd) MarshalTDAQ() ([]byte, error) {
-	buf := new(bytes.Buffer)
-	enc := NewEncoder(buf)
-	enc.WriteStr(cmd.Name)
-	enc.WriteStr(cmd.Addr)
-	return buf.Bytes(), enc.err
-}
-
-func (cmd *LogCmd) UnmarshalTDAQ(p []byte) error {
-	dec := NewDecoder(bytes.NewReader(p))
-	cmd.Name = dec.ReadStr()
-	cmd.Addr = dec.ReadStr()
-	return dec.err
-}
-
 var (
 	_ Cmder       = (*JoinCmd)(nil)
 	_ Marshaler   = (*JoinCmd)(nil)
@@ -403,10 +358,6 @@ var (
 	_ Cmder       = (*HBeatCmd)(nil)
 	_ Marshaler   = (*HBeatCmd)(nil)
 	_ Unmarshaler = (*HBeatCmd)(nil)
-
-	_ Cmder       = (*LogCmd)(nil)
-	_ Marshaler   = (*LogCmd)(nil)
-	_ Unmarshaler = (*LogCmd)(nil)
 
 	_ Cmder       = (*ConfigCmd)(nil)
 	_ Marshaler   = (*ConfigCmd)(nil)