--- a/flags/flag.go
+++ b/flags/flag.go
@@ -26,7 +26,7 @@ func New() config.Process {
 
 	flag.StringVar(&cmd.Name, "id", "", "name of the tdaq process")
 	flag.StringVar(&lvl, "lvl", "INFO", "msgstream level")
-	flag.StringVar(&cmd.Net, "net", "tcp", "network medium to use (tcp, unix) for data transfer")
+	flag.StringVar(&cmd.Trans, "net", "tcp", "network medium to use (tcp, unix) for data transfer")
 	flag.StringVar(&cmd.RunCtl, "rc-addr", ":44000", "[addr]:port of run-control process")
 
 	flag.Parse()
@@ -55,7 +55,7 @@ func NewRunControl() config.RunCtl {
 	flag.StringVar(&cmd.Name, "id", "", "name of the tdaq process")
 	flag.StringVar(&lvl, "lvl", "INFO", "msgstream level")
 	flag.StringVar(&cmd.RunCtl, "rc-addr", ":44000", "[addr]:port of run-ctl cmd server")
-	flag.StringVar(&cmd.Net, "net", "tcp", "network medium to use (tcp, unix) for data transfer")
+	flag.StringVar(&cmd.Trans, "net", "tcp", "network medium to use (tcp, unix) for data transfer")
 	flag.StringVar(&cmd.Web, "web", "", "[addr]:port of run-ctl web server")
 	flag.BoolVar(&cmd.Interactive, "i", false, "enable interactive run-ctl shell")