--- a/go.mod
+++ b/go.mod
@@ -3,11 +3,14 @@ module github.com/go-daq/tdaq
 go 1.13
 
 require (
+	github.com/Microsoft/go-winio v0.4.14 // indirect
 	github.com/mattn/go-runewidth v0.0.8 // indirect
 	github.com/peterh/liner v1.1.0
+	go.nanomsg.org/mangos/v3 v3.0.0
 	golang.org/x/exp v0.0.0-20200119233911-0405dc783f0a
 	golang.org/x/net v0.0.0-20200114155413-6afb5195e5aa
 	golang.org/x/sync v0.0.0-20190911185100-cd5d95a43a6e
+	golang.org/x/sys v0.0.0-20191010194322-b09406accb47 // indirect
 	golang.org/x/xerrors v0.0.0-20191204190536-9bdfabe68543
 	gonum.org/v1/gonum v0.6.2
 	gonum.org/v1/netlib v0.0.0-20191229114700-bbb4dff026f8 // indirect