--- a/internal/iomux/iomux.go
+++ b/internal/iomux/iomux.go
@@ -8,6 +8,7 @@ package iomux // import "github.com/go-daq/tdaq/internal/iomux"
 import (
 	"fmt"
 	"io"
+	"net"
 	"strings"
 	"sync"
 )
@@ -37,6 +38,42 @@ func (w *Writer) String() string {
 	return o.String()
 }
 
+func (w *Writer) Close() error {
+	w.mu.Lock()
+	defer w.mu.Unlock()
+	var err error
+	if wc, ok := w.w.(io.Closer); ok {
+		err = wc.Close()
+	}
+	return err
+}
+
 var (
 	_ io.Writer = (*Writer)(nil)
 )
+
+type Socket struct {
+	buf net.Buffers
+}
+
+func (sck *Socket) Write(p []byte) (int, error) {
+	sck.buf = append(sck.buf, p)
+	return len(p), nil
+}
+
+func (sck *Socket) Read(p []byte) (int, error) {
+	return sck.buf.Read(p)
+}
+
+func (sck *Socket) Send(p []byte) error {
+	sck.buf = append(sck.buf, p)
+	return nil
+}
+
+func (sck *Socket) Recv() ([]byte, error) {
+	n := len(sck.buf)
+	buf := sck.buf[n-1]
+	sck.buf = sck.buf[:n]
+
+	return buf, nil
+}