Fixes go-daq/tdaq#45.

```
    $> go test -run=NONE -bench=. -benchmem -v -count=30 ./xdaq |& tee new.txt
    goos: linux
    goarch: amd64
    pkg: github.com/go-daq/tdaq/xdaq
    BenchmarkSequence
    BenchmarkSequence-8            1        5002028033 ns/op            202809 bench-01-tokens/op/s     286290 bench-02-tof-ns          324556 bench-03-stddev-ns   337071736 B/op  13372408 allocs/op
    BenchmarkSequence-8            1        5001969676 ns/op            195844 bench-01-tokens/op/s     256068 bench-02-tof-ns          304096 bench-03-stddev-ns   325846184 B/op  12937195 allocs/op
    BenchmarkSequence-8            1        5002044200 ns/op            195556 bench-01-tokens/op/s     297051 bench-02-tof-ns          522221 bench-03-stddev-ns   324656240 B/op  12905361 allocs/op
    BenchmarkSequence-8            1        5002518027 ns/op            189894 bench-01-tokens/op/s     267915 bench-02-tof-ns          364368 bench-03-stddev-ns   316781816 B/op  12550672 allocs/op
    BenchmarkSequence-8            1        5004022773 ns/op            200847 bench-01-tokens/op/s     278057 bench-02-tof-ns          303669 bench-03-stddev-ns   331446312 B/op  13240768 allocs/op
    BenchmarkSequence-8            1        5002054559 ns/op            203135 bench-01-tokens/op/s     280511 bench-02-tof-ns          302128 bench-03-stddev-ns   334055824 B/op  13366392 allocs/op
    BenchmarkSequence-8            1        5001837162 ns/op            196953 bench-01-tokens/op/s     240237 bench-02-tof-ns          285960 bench-03-stddev-ns   324384080 B/op  12929206 allocs/op
    BenchmarkSequence-8            1        5002040677 ns/op            201544 bench-01-tokens/op/s     287401 bench-02-tof-ns          297372 bench-03-stddev-ns   333687864 B/op  13347889 allocs/op
    BenchmarkSequence-8            1        5002083382 ns/op            207074 bench-01-tokens/op/s     293068 bench-02-tof-ns          370191 bench-03-stddev-ns   340529160 B/op  13670303 allocs/op
    BenchmarkSequence-8            1        5002312164 ns/op            202941 bench-01-tokens/op/s    1860784 bench-02-tof-ns         7895683 bench-03-stddev-ns   338399976 B/op  13542604 allocs/op
    BenchmarkSequence-8            1        5002022532 ns/op            201617 bench-01-tokens/op/s     263770 bench-02-tof-ns          263538 bench-03-stddev-ns   332550112 B/op  13304645 allocs/op
    BenchmarkSequence-8            1        5001900673 ns/op            202157 bench-01-tokens/op/s    1813430 bench-02-tof-ns         6204176 bench-03-stddev-ns   333533600 B/op  13346478 allocs/op
    BenchmarkSequence-8            1        5001926190 ns/op            203433 bench-01-tokens/op/s     226467 bench-02-tof-ns          202620 bench-03-stddev-ns   332882128 B/op  13330889 allocs/op
    BenchmarkSequence-8            1        5001943500 ns/op            197872 bench-01-tokens/op/s     240577 bench-02-tof-ns          265023 bench-03-stddev-ns   325929704 B/op  13006945 allocs/op
    BenchmarkSequence-8            1        5001986002 ns/op            203323 bench-01-tokens/op/s     229970 bench-02-tof-ns          217928 bench-03-stddev-ns   332500872 B/op  13316182 allocs/op
    BenchmarkSequence-8            1        5001972356 ns/op            201984 bench-01-tokens/op/s     240948 bench-02-tof-ns          236297 bench-03-stddev-ns   330646056 B/op  13231726 allocs/op
    BenchmarkSequence-8            1        5001908190 ns/op            200533 bench-01-tokens/op/s     258893 bench-02-tof-ns          238778 bench-03-stddev-ns   329340568 B/op  13166889 allocs/op
    BenchmarkSequence-8            1        5003160682 ns/op            197904 bench-01-tokens/op/s     307223 bench-02-tof-ns          484487 bench-03-stddev-ns   325947544 B/op  13007606 allocs/op
    BenchmarkSequence-8            1        5001873473 ns/op            201747 bench-01-tokens/op/s     222933 bench-02-tof-ns          221429 bench-03-stddev-ns   330229592 B/op  13211426 allocs/op
    BenchmarkSequence-8            1        5002013202 ns/op            202396 bench-01-tokens/op/s     207532 bench-02-tof-ns          192325 bench-03-stddev-ns   330771264 B/op  13238783 allocs/op
    BenchmarkSequence-8            1        5001884360 ns/op            203672 bench-01-tokens/op/s     255260 bench-02-tof-ns          232172 bench-03-stddev-ns   334120080 B/op  13388145 allocs/op
    BenchmarkSequence-8            1        5001906341 ns/op            195245 bench-01-tokens/op/s     316223 bench-02-tof-ns          393334 bench-03-stddev-ns   325328896 B/op  12966099 allocs/op
    BenchmarkSequence-8            1        5002339936 ns/op            202943 bench-01-tokens/op/s     365559 bench-02-tof-ns          343830 bench-03-stddev-ns   339433360 B/op  13602720 allocs/op
    BenchmarkSequence-8            1        5001989522 ns/op            204002 bench-01-tokens/op/s     284346 bench-02-tof-ns          268818 bench-03-stddev-ns   336454400 B/op  13487320 allocs/op
    BenchmarkSequence-8            1        5002100357 ns/op            205693 bench-01-tokens/op/s     232281 bench-02-tof-ns          224703 bench-03-stddev-ns   335732400 B/op  13470733 allocs/op
    BenchmarkSequence-8            1        5002299509 ns/op            195242 bench-01-tokens/op/s     291469 bench-02-tof-ns          432107 bench-03-stddev-ns   324815760 B/op  12944743 allocs/op
    BenchmarkSequence-8            1        5001916066 ns/op            200065 bench-01-tokens/op/s     331542 bench-02-tof-ns         1485798 bench-03-stddev-ns   330273088 B/op  13196074 allocs/op
    BenchmarkSequence-8            1        5001974213 ns/op            200595 bench-01-tokens/op/s     886647 bench-02-tof-ns         3399154 bench-03-stddev-ns   328373136 B/op  13128903 allocs/op
    BenchmarkSequence-8            1        5001965016 ns/op            198701 bench-01-tokens/op/s     255577 bench-02-tof-ns          385490 bench-03-stddev-ns   326819008 B/op  13052250 allocs/op
    BenchmarkSequence-8            1        5001953632 ns/op            204182 bench-01-tokens/op/s     252710 bench-02-tof-ns          295482 bench-03-stddev-ns   333806368 B/op  13378933 allocs/op
    PASS
    ok      github.com/go-daq/tdaq/xdaq     150.313s
    
    $> benchstat ./ref.txt ./new.txt
    name        old time/op               new time/op               delta
    Sequence-8                5.00s ± 0%                5.00s ± 0%     +0.02%  (p=0.000 n=26+26)
    
    name        old bench-01-tokens/op/s  new bench-01-tokens/op/s  delta
    Sequence-8                39.2k ± 5%               201.0k ± 3%   +412.51%  (p=0.000 n=30+29)
    
    name        old bench-02-tof-ns       new bench-02-tof-ns       delta
    Sequence-8                18.2k ±18%               269.3k ±36%  +1376.06%  (p=0.000 n=25+27)
    
    name        old bench-03-stddev-ns    new bench-03-stddev-ns    delta
    Sequence-8               54.3k ±174%               306.7k ±70%   +465.12%  (p=0.000 n=25+26)
    
    name        old alloc/op              new alloc/op              delta
    Sequence-8                102MB ± 4%                331MB ± 4%   +224.56%  (p=0.000 n=29+30)
    
    name        old allocs/op             new allocs/op             delta
    Sequence-8                3.34M ± 5%               13.22M ± 5%   +296.38%  (p=0.000 n=30+30)
```